+++ src/main/java/com/idealista/infrastructure/api/inbound/AdsController.java
@@ -1,9 +1,11 @@
-package com.idealista.infrastructure.api;
+package com.idealista.infrastructure.api.inbound;
 
-import java.util.List;
-
+import com.idealista.infrastructure.api.inbound.dto.PublicAd;
+import com.idealista.infrastructure.api.inbound.dto.QualityAd;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 public class AdsController {
@@ -16,12 +18,6 @@ public class AdsController {
 
     //TODO añade url del endpoint
     public ResponseEntity<List<PublicAd>> publicListing() {
-        //TODO rellena el cuerpo del método
-        return ResponseEntity.notFound().build();
-    }
-
-    //TODO añade url del endpoint
-    public ResponseEntity<Void> calculateScore() {
         //TODO rellena el cuerpo del método
         return ResponseEntity.notFound().build();
     }
