+++ src/main/java/com/idealista/infrastructure/api/inbound/dto/PublicAd.java
@@ -1,4 +1,4 @@
-package com.idealista.infrastructure.api;
+package com.idealista.infrastructure.api.inbound.dto;
 
 import java.util.List;
